+++ test/td_audit/notifications_test.exs
@@ -80,7 +80,7 @@ defmodule TdAudit.NotificationsTest do
               %{
                 max_event_id: ^event_id,
                 notifications: _created_notification_ids,
-                no_subcription_events: [%{id: ^event_id}],
+                no_subscription_events: [%{id: ^event_id}],
                 no_subscription_events_recipient_ids: %{^event_id => [1]},
               }} = Notifications.create(periodicity: "minutely")
     end